third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDiffuseLightingElement-dom-surfaceScale-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'surfaceScale' attribute of the SVGFEDiffuseLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.getAttribute('surfaceScale') is "0"
PASS gradientElement.getAttribute('surfaceScale') is "1"
PASS successfullyParsed is true

TEST COMPLETE
